+++ business/src/main/java/com/abdecd/moebackend/business/pojo/vo/backstage/commonVideoGroup/VideoGroupVO.java
@@ -1,7 +1,7 @@
 package com.abdecd.moebackend.business.pojo.vo.backstage.commonVideoGroup;
 
 import com.abdecd.moebackend.business.dao.entity.VideoGroupTag;
-import com.abdecd.moebackend.business.pojo.vo.common.UploaderVO;
+import com.abdecd.moebackend.business.pojo.vo.plainuser.UploaderVO;
 import lombok.Data;
 import lombok.experimental.Accessors;
 
@@ -16,6 +16,6 @@ public class VideoGroupVO {
     private String cover;
     private UploaderVO uploader;
     private ArrayList<VideoGroupTag> tags;
-    private Integer type;
+    private Byte type;
 }
